feat: update directory schedule and subscription plan entities
