+++ apps/api/.env.example
@@ -215,10 +215,9 @@ SMTP_SECURE=false
 SMTP_IGNORE_TLS=false
 
 # =============================================================================
-# TRIGGER.DEV CONFIGURATION
+# TWENTY CONFIGURATION
 # =============================================================================
 
-TRIGGER_SECRET_KEY=
 TWENTY_CRM_BASE_URL=https://api.twenty.com
 TWENTY_CRM_API_KEY= no 
 TWENTY_CRM_WORKSPACE_ID=
