+++ package.json
@@ -44,7 +44,7 @@
 		"node": ">=20",
 		"pnpm": ">=9.9.0"
 	},
-	"packageManager": "pnpm@10.22.0",
+	"packageManager": "pnpm@10.23.0",
 	"prettier": {
 		"printWidth": 120,
 		"singleQuote": true,
