+++ packages/agent/src/database/repositories/subscription-plan.repository.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { SubscriptionPlan, SubscriptionPlanCode } from '@src/entities/subscription-plan.entity';
+import { SubscriptionPlan, SubscriptionPlanCode } from '@src/entities';
 
 @Injectable()
 export class SubscriptionPlanRepository {
