+++ packages/agent/src/entities/types.ts
@@ -14,3 +14,28 @@ export type GenerateStatus = {
     step?: string;
     error?: string;
 };
+
+export enum DirectoryScheduleCadence {
+    HOURLY = 'hourly',
+    DAILY = 'daily',
+    WEEKLY = 'weekly',
+    MONTHLY = 'monthly',
+}
+
+export enum DirectoryScheduleStatus {
+    DISABLED = 'disabled',
+    ACTIVE = 'active',
+    PAUSED = 'paused',
+    CANCELED = 'canceled',
+}
+
+export enum DirectoryScheduleBillingMode {
+    SUBSCRIPTION = 'subscription',
+    USAGE = 'usage',
+}
+
+export enum SubscriptionPlanCode {
+    FREE = 'free',
+    STANDARD = 'standard',
+    PREMIUM = 'premium',
+}
