+++ packages/agent/src/services/directory-schedule.service.ts
@@ -1,10 +1,5 @@
 import { BadRequestException, Injectable, Logger, NotFoundException } from '@nestjs/common';
-import {
-    DirectorySchedule,
-    DirectoryScheduleBillingMode,
-    DirectoryScheduleCadence,
-    DirectoryScheduleStatus,
-} from '@src/entities/directory-schedule.entity';
+import { DirectorySchedule } from '@src/entities/directory-schedule.entity';
 import { DirectoryScheduleRepository } from '@src/database/repositories/directory-schedule.repository';
 import { DirectoryRepository } from '@src/database/repositories/directory.repository';
 import { DirectoryOwnershipService } from './directory-ownership.service';
@@ -16,7 +11,12 @@ import {
 } from '@src/dto';
 import { User } from '@src/entities/user.entity';
 import { config } from '@src/config';
-import { GenerateStatusType } from '@src/entities/types';
+import {
+    DirectoryScheduleBillingMode,
+    DirectoryScheduleCadence,
+    DirectoryScheduleStatus,
+    GenerateStatusType,
+} from '@src/entities/types';
 import { UsageLedgerService } from '@src/subscriptions/usage-ledger.service';
 import { UsageLedgerTriggerType } from '@src/entities/usage-ledger-entry.entity';
